+++ src/core/context/ArticleCommentsProvider.tsx
@@ -1,7 +1,5 @@
-import { QueryFunctionContext, useInfiniteQuery } from "@tanstack/react-query";
-import { useRouter } from "next/router";
-import { createContext, ReactNode, useContext, useMemo, useState } from "react";
-import getArticleComments from "../api/boards/getArticleComments";
+import { createContext, ReactNode, useContext, useMemo } from "react";
+import useArticleCommentsQuery from "@/lib/hooks/article/useArticleCommentsQuery";
 import { ArticleComment } from "../dtos/boards/boards";
 
 interface ArticleCommentsContextValues {
@@ -17,10 +15,6 @@ interface ArticleCommentsContextValues {
   refetchComments: () => void;
 }
 
-interface InfiniteQueryContext extends QueryFunctionContext {
-  pageParam?: number | null;
-}
-
 const initialContextValues: ArticleCommentsContextValues = {
   commentsCount: undefined,
   increaseCommentsCount: () => {},
@@ -34,8 +28,6 @@ const initialContextValues: ArticleCommentsContextValues = {
   refetchComments: () => {},
 };
 
-const LIMIT = 5;
-
 const ArticleCommentsContext = createContext(initialContextValues);
 
 export function ArticleCommentsProvider({
@@ -45,50 +37,21 @@ export function ArticleCommentsProvider({
   initialCommentsCount: number;
   children: ReactNode;
 }) {
-  const { query } = useRouter();
-  const articleId = query.id as string;
-
-  const [commentsCount, setCommentsCount] = useState(initialCommentsCount);
-
   const {
-    data: commentsData,
+    commentsCount,
+    increaseCommentsCount,
+    decreaseCommentsCount,
+    comments,
     hasNextPage,
-    fetchNextPage,
-    isLoading: isCommentsLoading,
-    isFetching: isFetchingComments,
-    isError: isCommentsError,
-    refetch,
-  } = useInfiniteQuery({
-    queryKey: ["ArticleComments", articleId],
-    queryFn: ({ pageParam }: InfiniteQueryContext) =>
-      getArticleComments({
-        articleId,
-        limit: LIMIT,
-        cursor: pageParam ?? undefined,
-      }),
-    initialPageParam: undefined,
-    getNextPageParam: (lastPage) => lastPage.nextCursor,
-    staleTime: 1000 * 60,
-    gcTime: 0,
-  });
+    fetchNextComments,
+    isCommentsLoading,
+    isFetchingComments,
+    isCommentsError,
+    refetchComments,
+  } = useArticleCommentsQuery(initialCommentsCount);
 
-  const comments = commentsData?.pages.map((e) => e.list).flat();
-
-  const contextValues = useMemo(() => {
-    const increaseCommentsCount = () => {
-      setCommentsCount((prev) => prev + 1);
-    };
-
-    const decreaseCommentsCount = () => {
-      setCommentsCount((prev) => prev - 1);
-    };
-    const fetchNextComments = () => {
-      fetchNextPage();
-    };
-    const refetchComments = () => {
-      refetch();
-    };
-    return {
+  const contextValues = useMemo(
+    () => ({
       commentsCount,
       increaseCommentsCount,
       decreaseCommentsCount,
@@ -99,17 +62,18 @@ export function ArticleCommentsProvider({
       isFetchingComments,
       isCommentsError,
       refetchComments,
-    };
-  }, [
-    commentsCount,
-    comments,
-    hasNextPage,
-    fetchNextPage,
-    isCommentsLoading,
-    isFetchingComments,
-    isCommentsError,
-    refetch,
-  ]);
+    }),
+    [
+      commentsCount,
+      comments,
+      hasNextPage,
+      fetchNextComments,
+      isCommentsLoading,
+      isFetchingComments,
+      isCommentsError,
+      refetchComments,
+    ],
+  );
 
   return (
     <ArticleCommentsContext.Provider value={contextValues}>
